refactor(redirect): single-hop loop check, ban '..', centralise constants
- Simplify getChainedRedirect to a single hop: drop the 10-hop walk and
  visited-set. The mirror pair (/a->/b while /b->/a) is the documented loop
  scope; longer cycles are intentionally not chased. This also makes the loop
  description accurate (the destination really does redirect back to source).
- Reject '..' path segments in any destination, not just internal paths.
- Move RedirectValidationCode, REDIRECT_MESSAGES, RESERVED_SOURCE_PREFIXES and
  the validation-result contract to ~/constants/redirect.
